+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,11 +1,13 @@
 object Versions {
     const val KOTLIN = "1.6.10"
+    const val ATOMIC_FU = "0.17.0"
     const val ANDROID_GRADLE_PLUGIN = "7.1.2"
     const val JETPACK_COMPOSE = "1.2.0-alpha03"
     const val COIL = "1.4.0"
     const val KTLINT = "10.2.0"
     const val KOTLINX_SERIALIZATION = "1.3.2"
-    const val KTOR = "1.6.7"
+    const val KOTLINX_COROUTINES = "1.6.0"
+    const val KTOR = "2.0.0-beta-1"
     const val KOTLIN_WRAPPERS_EXTENSIONS = "1.0.1-pre.290"
     const val ANDROIDX_LIFECYCLE = "2.4.0"